Move LogApiOptions.h to RestServer
